Renamed xml renderer to svg-headless
