+++ src/render/svg-headless/Renderer.js
