+++ src/render/svg-headless/index.js
