+++ src/render/svg-headless/svg.js
